--- pkg/services/sqlstore/login_attempt.go
@@ -1,1 +1,0 @@
-package sqlstore
