--- pkg/services/sqlstore/login_attempt_test.go
@@ -1,1 +1,0 @@
-package sqlstore
