+++ pkg/services/sqlstore/store.go
@@ -62,12 +62,6 @@ type Store interface {
 	Sync() error
 	Reset() error
 	Quote(value string) string
-	UpdateTempUserStatus(ctx context.Context, cmd *models.UpdateTempUserStatusCommand) error
-	CreateTempUser(ctx context.Context, cmd *models.CreateTempUserCommand) error
-	UpdateTempUserWithEmailSent(ctx context.Context, cmd *models.UpdateTempUserWithEmailSentCommand) error
-	GetTempUsersQuery(ctx context.Context, query *models.GetTempUsersQuery) error
-	GetTempUserByCode(ctx context.Context, query *models.GetTempUserByCodeQuery) error
-	ExpireOldUserInvites(ctx context.Context, cmd *models.ExpireTempUsersCommand) error
 	GetDBHealthQuery(ctx context.Context, query *models.GetDBHealthQuery) error
 	IsAdminOfTeams(ctx context.Context, query *models.IsAdminOfTeamsQuery) error
 	GetSqlxSession() *session.SessionDB
