+++ pkg/services/temp_user/tempuserimpl/temp_user.go
@@ -4,25 +4,24 @@ import (
 	"context"
 
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/sqlstore"
+	"github.com/grafana/grafana/pkg/services/sqlstore/db"
 	tempuser "github.com/grafana/grafana/pkg/services/temp_user"
 )
 
 type Service struct {
-	// TODO remove sqlstore
-	sqlStore *sqlstore.SQLStore
+	store store
 }
 
 func ProvideService(
-	ss *sqlstore.SQLStore,
+	db db.DB,
 ) tempuser.Service {
 	return &Service{
-		sqlStore: ss,
+		store: &xormStore{db: db},
 	}
 }
 
 func (s *Service) UpdateTempUserStatus(ctx context.Context, cmd *models.UpdateTempUserStatusCommand) error {
-	err := s.sqlStore.UpdateTempUserStatus(ctx, cmd)
+	err := s.store.UpdateTempUserStatus(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -30,7 +29,7 @@ func (s *Service) UpdateTempUserStatus(ctx context.Context, cmd *models.UpdateTe
 }
 
 func (s *Service) CreateTempUser(ctx context.Context, cmd *models.CreateTempUserCommand) error {
-	err := s.sqlStore.CreateTempUser(ctx, cmd)
+	err := s.store.CreateTempUser(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -38,7 +37,7 @@ func (s *Service) CreateTempUser(ctx context.Context, cmd *models.CreateTempUser
 }
 
 func (s *Service) UpdateTempUserWithEmailSent(ctx context.Context, cmd *models.UpdateTempUserWithEmailSentCommand) error {
-	err := s.sqlStore.UpdateTempUserWithEmailSent(ctx, cmd)
+	err := s.store.UpdateTempUserWithEmailSent(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -46,7 +45,7 @@ func (s *Service) UpdateTempUserWithEmailSent(ctx context.Context, cmd *models.U
 }
 
 func (s *Service) GetTempUsersQuery(ctx context.Context, cmd *models.GetTempUsersQuery) error {
-	err := s.sqlStore.GetTempUsersQuery(ctx, cmd)
+	err := s.store.GetTempUsersQuery(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -54,7 +53,7 @@ func (s *Service) GetTempUsersQuery(ctx context.Context, cmd *models.GetTempUser
 }
 
 func (s *Service) GetTempUserByCode(ctx context.Context, cmd *models.GetTempUserByCodeQuery) error {
-	err := s.sqlStore.GetTempUserByCode(ctx, cmd)
+	err := s.store.GetTempUserByCode(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -62,7 +61,7 @@ func (s *Service) GetTempUserByCode(ctx context.Context, cmd *models.GetTempUser
 }
 
 func (s *Service) ExpireOldUserInvites(ctx context.Context, cmd *models.ExpireTempUsersCommand) error {
-	err := s.sqlStore.ExpireOldUserInvites(ctx, cmd)
+	err := s.store.ExpireOldUserInvites(ctx, cmd)
 	if err != nil {
 		return err
 	}
